refactor: add useSpeechRecogntion separated from WriteVoice
